@@ -13,21 +13,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.0",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "~12.2.0",
-    "@angular/compiler": "~12.2.0",
-    "@angular/core": "~12.2.0",
-    "@angular/forms": "~12.2.0",
-    "@angular/platform-browser": "~12.2.0",
-    "@angular/platform-browser-dynamic": "~12.2.0",
-    "@angular/router": "~12.2.0",
+    "@angular/animations": "~13.3.11",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "~13.3.11",
+    "@angular/compiler": "~13.3.11",
+    "@angular/core": "~13.3.11",
+    "@angular/forms": "~13.3.11",
+    "@angular/platform-browser": "~13.3.11",
+    "@angular/platform-browser-dynamic": "~13.3.11",
+    "@angular/router": "~13.3.11",
     "@ctrl/ngx-emoji-mart": "^6.2.0",
     "@microsoft/signalr": "^6.0.4",
     "@ngx-pwa/local-storage": "^13.0.2",
     "date-fns": "^2.28.0",
     "echarts": "^5.2.2",
-    "rxjs": "~6.6.0",
     "tds-report": "https://tang-ui.tpos.dev/lib/tds-report-1.1.0.tgz",
     "tmt-tang-ui": "https://tang-ui.tpos.dev/lib/tmt-tang-ui-1.3.0.tgz",
     "tslib": "^2.3.0",
@@ -35,9 +34,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.16",
-    "@angular/cli": "~12.2.16",
-    "@angular/compiler-cli": "~12.2.0",
+    "@angular-devkit/build-angular": "~13.3.7",
+    "@angular/cli": "~13.3.7",
+    "@angular/compiler-cli": "~13.3.11",
     "@tailwindcss/forms": "^0.5.0",
     "@tailwindcss/typography": "^0.5.2",
     "@types/jasmine": "~3.8.0",
@@ -51,7 +50,8 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
     "postcss": "^8.4.7",
+    "rxjs": "^7.5.5",
     "tailwindcss": "^3.0.23",
-    "typescript": "~4.3.5"
+    "typescript": "~4.6.4"
   }
 }
