@@ -1,8 +1,8 @@
-import { Component, OnChanges, OnDestroy, OnInit, Optional, Host, SkipSelf } from '@angular/core';
-import { Observable, Subject, Subscription } from 'rxjs';
+import { Component, OnDestroy, OnInit, Optional, Host, SkipSelf } from '@angular/core';
+import { Observable, Subject } from 'rxjs';
 import { CRMTeamDTO } from 'src/app/main-app/dto/team/team.dto';
 import { ConversationPostFacade } from 'src/app/main-app/services/facades/conversation-post.facade';
-import { TDSHelperArray, TDSHelperObject, TDSHelperString, TDSMessageService } from 'tmt-tang-ui';
+import { TDSHelperArray, TDSHelperObject } from 'tmt-tang-ui';
 import { CommentByPost } from 'src/app/main-app/dto/conversation/post/comment-post.dto';
 import { ItemPostCommentComponent } from '../../conversation-post/components/item-post-comment/item-post-comment.component';
 
@@ -21,7 +21,7 @@ export class ManagePostCommentComponent implements OnInit, OnDestroy {
   otherSelecteds: any = [];
   partners$!: Observable<any>;
 
-  destroy$ = new Subject();
+  destroy$ = new Subject<void>();
   messageModel!: string;
 
   constructor(private conversationPostFacade: ConversationPostFacade,
