@@ -33,7 +33,7 @@ export class PostCommentAllComponent implements OnInit, OnDestroy {
   childs: any = {};
   commentOrders: any = [];
   partners$!: Observable<any>;
-  destroy$ = new Subject();
+  destroy$ = new Subject<void>();
   messageModel!: string;
   currentId:string = '';
 
