@@ -17,7 +17,7 @@ export class ConversationPostFacade extends BaseSevice implements OnDestroy {
   prefix: string = "";
   table: string = "";
   baseRestApi: string = "";
-  destroy$ = new Subject();
+  destroy$ = new Subject<void>();
 
   public from: any;
   public comment: any;
@@ -48,7 +48,7 @@ export class ConversationPostFacade extends BaseSevice implements OnDestroy {
         this.from = Object.assign({}, res.from);
         this.partner = null;
         // TODO: Tải khách hàng theo id người dùng bình luận
-        
+
       } else {
         // Thông tin cũ, không làm gì cả
       }
