@@ -4,7 +4,7 @@ import { map } from 'rxjs/operators';
 
 export class NetworkHelper {
   public static checkNetwork(): Observable<any> {
-    return merge<boolean>(
+    return merge<any>(
       fromEvent(window, 'offline').pipe(map(() => false)),
       fromEvent(window, 'online').pipe(map(() => true)),
       new Observable((sub: Observer<boolean>) => {
